+++ src/elements/conditions/OwnerSlugConditionRule.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace benf\neo\elements\conditions;
+
+use Craft;
+use craft\base\ElementInterface;
+use craft\elements\conditions\SlugConditionRule;
+
+/**
+ * Class OwnerSlugConditionRule
+ *
+ * @package benf\neo\elements\conditions
+ * @author Spicy Web <plugins@spicyweb.com.au>
+ * @since 5.5.0
+ */
+class OwnerSlugConditionRule extends SlugConditionRule
+{
+    use OwnerConditionRuleTrait;
+
+    /**
+     * @inheritdoc
+     */
+    public function getLabel(): string
+    {
+        return Craft::t('neo', 'Owner Slug');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function matchElement(ElementInterface $element): bool
+    {
+        return $this->_matchElement($element);
+    }
+}